--- lib/widgets/bottombaritem.dart
@@ -1,1 +1,0 @@
-class Config {}
